use new docusaurus 'fast build' and (hopefully) improve CI time w/conditional dependency installs