@rcoder 2024-12-20
before:

`yarn build  72.10s user 5.94s system 278% cpu 28.068 total`

after `future.experimental_faster`:

`yarn build  17.70s user 2.02s system 280% cpu 7.035 total`